generate indexes.
