+++ kitsune/search/tests/__init__.py
@@ -1,13 +1,10 @@
-from django.conf import settings
 from django.test.client import RequestFactory
 from django.test.utils import override_settings
 
 import factory
-from elasticutils.contrib.django import get_es
 
-from kitsune.search import es_utils
-from kitsune.search.models import generate_tasks, Synonym
-from kitsune.sumo.tests import SkipTest, TestCase
+from kitsune.search.models import Synonym
+from kitsune.sumo.tests import TestCase
 
 
 # Dummy request for passing to question_searcher() and brethren.
@@ -17,64 +14,7 @@ dummy_request = RequestFactory().get('/')
 @override_settings(ES_LIVE_INDEXING=True)
 class ElasticTestCase(TestCase):
     """Base class for Elastic Search tests, providing some conveniences"""
-    skipme = False
     search_tests = True
-
-    @classmethod
-    def setUpClass(cls):
-        super(ElasticTestCase, cls).setUpClass()
-
-        if not getattr(settings, 'ES_URLS'):
-            cls.skipme = True
-            return
-
-        # try to connect to ES and if it fails, skip ElasticTestCases.
-        if not get_es().ping():
-            cls.skipme = True
-            return
-
-    def setUp(self):
-        if self.skipme:
-            raise SkipTest
-
-        super(ElasticTestCase, self).setUp()
-        self.setup_indexes()
-
-    def tearDown(self):
-        super(ElasticTestCase, self).tearDown()
-        self.teardown_indexes()
-
-    def refresh(self, run_tasks=True):
-        es = get_es()
-
-        if run_tasks:
-            # Any time we're doing a refresh, we're making sure that
-            # the index is ready to be queried. Given that, it's
-            # almost always the case that we want to run all the
-            # generated tasks, then refresh.
-            generate_tasks()
-
-        for index in es_utils.all_write_indexes():
-            es.indices.refresh(index=index)
-
-        es.cluster.health(wait_for_status='yellow')
-
-    def reindex_and_refresh(self):
-        """Reindexes anything in the db"""
-        from kitsune.search.es_utils import es_reindex_cmd
-        es_reindex_cmd()
-        self.refresh(run_tasks=False)
-
-    def setup_indexes(self, empty=False, wait=True):
-        """(Re-)create write index"""
-        from kitsune.search.es_utils import recreate_indexes
-        recreate_indexes()
-        get_es().cluster.health(wait_for_status='yellow')
-
-    def teardown_indexes(self):
-        """Tear down write index"""
-        for index in es_utils.all_write_indexes():
-            es_utils.delete_index(index)
 
 
 class SynonymFactory(factory.DjangoModelFactory):
